Clean up folder structure for selectors
